+++ CHANGELOG.md
@@ -11,8 +11,8 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- **Breaking:** The ``solver`` keyword argument is now mandatory
 - Quadratic programming example now randomly selects an available solver
-- The ``solver`` kwargs is now mandatory (previously defaulted to "quadprog")
 
 ## [1.9.1] - 2022/05/02